pyproject.toml
[tool.poetry]
name = "ab-test-api"
version = "0.1.0"
description = ""
authors = ["Ilia Kaziamov <kaziamov@outlook.com>"]
readme = "README.md"
packages = [{include = "ab_test_api"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^1.0.0"
uvicorn = "^0.21.1"
pony = "^0.7.16"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
